feat-#16: move dtos to application
