+++ src/item/application/commands/impl/update-item.command.ts
@@ -1,4 +1,4 @@
-import { UpdateItemDto } from '../../../dtos';
+import { UpdateItemDto } from '../../dtos';
 
 export class UpdateItemCommand {
   constructor(public readonly updateItemDto: UpdateItemDto) {}
