+++ src/item/application/controllers/item.controller.ts
@@ -1,7 +1,7 @@
 import { Controller } from '@nestjs/common';
 import { MessagePattern } from '@nestjs/microservices';
 
-import { CreateItemDto, GetItemsDto, UpdateItemDto } from '../../dtos';
+import { CreateItemDto, GetItemsDto, UpdateItemDto } from '../dtos';
 import { ItemEntity } from '../../infrastructure/entities';
 import { ItemService } from '../services';
