+++ src/item/domain/events/impl/item-updated.event.ts
@@ -1,6 +1,6 @@
 import { IEvent } from '@nestjs/cqrs';
 
-import { UpdateItemDto } from '../../../dtos';
+import { UpdateItemDto } from '../../../application/dtos';
 
 export class ItemUpdatedEvent implements IEvent {
   constructor(public readonly updateItemDto: UpdateItemDto) {}
