+++ src/item/domain/models/item.model.ts
@@ -1,6 +1,6 @@
 import { AggregateRoot } from '@nestjs/cqrs';
 
-import { CreateItemDto, UpdateItemDto } from '../../dtos';
+import { CreateItemDto, UpdateItemDto } from '../../application/dtos';
 import {
   ItemCreatedEvent,
   ItemDeletedEvent,
