+++ src/item/domain/repositories/item.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
-import { CreateItemDto, UpdateItemDto } from '../../dtos';
+import { CreateItemDto, UpdateItemDto } from '../../application/dtos';
 import { Item } from '../models';
 import { ItemEventType } from '../types';
